optimize logger
